requirements.txt
# Core Framework
langchain==0.1.0
langchain-community==0.0.10

# Hugging Face - Using Inference API (no local model downloads)
transformers==4.36.2
huggingface-hub==0.20.2
sentence-transformers==2.3.1

# UI Framework
gradio==4.13.0

# Vector Store
faiss-cpu==1.7.4

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
Pillow==10.2.0
python-pptx==0.6.23

# NLP & Language Processing
spacy==3.7.2
langdetect==1.0.9

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1

# Graph Visualization
networkx==3.2.1
matplotlib==3.8.2

# Additional Tools
pytesseract==0.3.10
openpyxl==3.1.2
